reworkFolderStructure
