+++ client/src/app/coach/page.tsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import {Container} from "react-bootstrap";
+
+function Page() {
+    return (
+        <Container>
+coach
+        </Container>
+    );
+}
+
+export default Page;
