+++ client/src/app/diet/page.tsx
@@ -1,0 +1,10 @@
+import React from 'react';
+import DietTabs from "@/components/Diet/DietTabs";
+
+function Page() {
+    return (
+        <DietTabs />
+    );
+}
+
+export default Page;
