+++ client/src/app/layout.tsx
@@ -1,10 +1,11 @@
 import type {Metadata} from 'next'
 import {Inter} from 'next/font/google'
-import 'bootstrap/dist/css/bootstrap.min.css'
-import './globals.css'
-import RootLayoutComponent from "@/components/RootLayout";
 import React from "react";
-import Link from "next/link";
+import Image from "next/image"
+import {Container, Nav, Navbar} from "react-bootstrap";
+import TopNavbar from "@/components/Navigation/TopNavbar";
+import "bootstrap/dist/css/bootstrap.css"
+import './globals.css'
 import {CookiesProvider} from "next-client-cookies/server";
 
 
@@ -12,7 +13,9 @@ const inter = Inter({subsets: ['latin']})
 
 export const metadata: Metadata = {
     title: 'THRIVE TRIBE',
-    description: 'Generated by create next app',
+    icons: {
+        icon: "/pictures/Logo.png"
+    }
 }
 
 export default function RootLayout({
@@ -23,22 +26,50 @@ export default function RootLayout({
     const linkStyle = {
         color: "#1E1E1E"
     };
-
     return (
-        <html lang="en">
-        <head>
-            <link href='https://unpkg.com/boxicons@2.1.4/css/boxicons.min.css' rel='stylesheet'/>
-            <link rel="icon" type="image/png" href="/pictures/Logo.png"/>
-            <title>Thrive Tribe</title>
-        </head>
+        <html>
         <body>
+        <CookiesProvider>
+            <TopNavbar/>
 
-        <RootLayoutComponent>
+            <Container fluid className={"p-0 m-0 d-flex justify-content-center"}>
+                {children}
+            </Container>
 
-            {children}
-
-        </RootLayoutComponent>
-
+            <footer className="footer">
+                <Container>
+                    <div className="top">
+                        <div className="logo">
+                            <Image alt="Logo" src="/pictures/Logo.png" width={50} height={50}></Image>
+                            THRIVE TRIBE
+                        </div>
+                        <ul>
+                            <li><a href="/">Home</a></li>
+                            <li><a href="#">Blog</a></li>
+                            <li><a href="#">Contact</a></li>
+                            <li><a href="#">Future</a></li>
+                        </ul>
+                        <div className="social-links">
+                            <a href="#"><i className='bx bxl-facebook-square'></i></a>
+                            <a href="#"><i className='bx bxl-gmail'></i></a>
+                            <a href="#"><i className='bx bxl-twitter'></i></a>
+                            <a href="#"><i className='bx bxl-linkedin-square'></i></a>
+                        </div>
+                    </div>
+                    <div className="separator"></div>
+                    <div className="bottom">
+                        <p>
+                            Made with Intentions by Thrive Tribe
+                        </p>
+                        <div className="links">
+                            <a href="#">Privacy Policy</a>
+                            <a href="#">Terms of Service</a>
+                            <a href="#">Cookies</a>
+                        </div>
+                    </div>
+                </Container>
+            </footer>
+        </CookiesProvider>
         </body>
         </html>
     )
