+++ client/src/app/login/page.tsx
@@ -1,0 +1,13 @@
+import React from 'react';
+import "./login.css";
+import LoginForm from "@/components/Login/LoginForm";
+
+function Page() {
+    return (
+        <div>
+            <LoginForm/>
+        </div>
+    );
+}
+
+export default Page;
