+++ client/src/app/messageboard/page.tsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import Datagrid from "@/components/DataGrid/Datagrid";
+
+function Page() {
+    return (
+        <>
+            <Datagrid/>
+        </>
+    );
+}
+
+export default Page;
