--- client/src/pages/login/index.tsx
@@ -1,12 +1,0 @@
-import React, {useState, useEffect} from 'react'
-import RootLayoutComponent from "@/components/RootLayout";
-import Login from "@/components/Login/Login";
-
-
-export default function LoginForm() {
-    return (
-        <RootLayoutComponent>
-            <Login/>
-        </RootLayoutComponent>
-    )
-}
