--- client/src/pages/messageboard/index.tsx
@@ -1,9 +1,0 @@
-import RootLayoutComponent from "@/components/RootLayout";
-
-export default function index(){
-    return(
-        <RootLayoutComponent>
-            <h1>DOLGOZZ, MERT AZT MONDJUK NEM CSINÁLTÁL SEMMIT!!!</h1>
-        </RootLayoutComponent>
-    )
-}
